fixed sponsors issue [adblock]
So this was the problem: stuff with a class of "sponsor-logo" is BLOCKED. ARE YOU KIDDING ME. IM NOT ADVERTISING. PLEASE. so, yeah, i'm a bit tilted after trying to figure out why it works for some people and not others, independent of their operating system, browser, resolution, or just about anything else. it was frickin' ad block. welp. if this ever is read by anyone, don't make this same mistake. god im sick of this